pre_onboarding/requirements.txt
asgiref==3.5.0
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
generics==2.0.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
